MNT: Remove install_requires from setup.py
This is so the conda package will build correctly.  Downloading with
setuptools is disabled during conda-builds, so anything that is listed
in the install_requires will try to be installed during the **build**
phase of the conda-recipe, but many of these are actually runtime
dependencies.  This causes the conda-build to fail.